+++ packages/terra-dropdown-button/tests/jest/DropdownList.test.jsx
@@ -1,16 +1,18 @@
 import React from 'react';
 import ThemeContextProvider from 'terra-theme-context/lib/ThemeContextProvider';
-
+/* eslint-disable-next-line import/no-extraneous-dependencies */
+import { shallowWithIntl, mountWithIntl } from 'terra-enzyme-intl';
+/* eslint-disable-next-line import/no-extraneous-dependencies */
 import DropdownList from '../../src/_DropdownList';
 import { Item } from '../../src/DropdownButton';
 
 describe('Dropdown List', () => {
   it('renders a default dropdown list', () => {
-    const wrapper = shallow(
+    const wrapper = shallowWithIntl(
       <DropdownList requestClose={() => {}}>
         <Item label="1st Option" onSelect={() => {}} />
       </DropdownList>,
-    );
+    ).dive();
 
     expect(wrapper).toMatchSnapshot();
     // ensures both data attributes are added when list contains single item
@@ -19,7 +21,7 @@ describe('Dropdown List', () => {
   });
 
   it('renders a dropdown list with a set width', () => {
-    const wrapper = shallow(
+    const wrapper = shallowWithIntl(
       <DropdownList requestClose={() => {}} width="440px">
         <Item label="1st Option" onSelect={() => {}} />
       </DropdownList>,
@@ -28,7 +30,7 @@ describe('Dropdown List', () => {
   });
 
   it('renders a dropdown list with multiple children', () => {
-    const wrapper = shallow(
+    const wrapper = shallowWithIntl(
       <DropdownList requestClose={() => {}}>
         <Item label="1st Option" onSelect={() => {}} />
         <Item label="2nd Option" onSelect={() => {}} />
@@ -39,7 +41,7 @@ describe('Dropdown List', () => {
   });
 
   it('renders a dropdown list a non-default focused option', () => {
-    const wrapper = shallow(
+    const wrapper = shallowWithIntl(
       <DropdownList requestClose={() => {}}>
         <Item label="1st Option" onSelect={() => {}} />
         <Item label="2nd Option" onSelect={() => {}} />
@@ -52,7 +54,7 @@ describe('Dropdown List', () => {
   });
 
   it('renders a dropdown list an active option', () => {
-    const wrapper = shallow(
+    const wrapper = shallowWithIntl(
       <DropdownList requestClose={() => {}}>
         <Item label="1st Option" onSelect={() => {}} />
         <Item label="2nd Option" onSelect={() => {}} />
@@ -65,7 +67,7 @@ describe('Dropdown List', () => {
   });
 
   it('renders a dropdown list an active and focused option', () => {
-    const wrapper = shallow(
+    const wrapper = shallowWithIntl(
       <DropdownList requestClose={() => {}}>
         <Item label="1st Option" onSelect={() => {}} />
         <Item label="2nd Option" onSelect={() => {}} />
@@ -78,7 +80,7 @@ describe('Dropdown List', () => {
   });
 
   it('correctly applies the theme context className', () => {
-    const wrapper = mount(
+    const wrapper = mountWithIntl(
       <ThemeContextProvider theme={{ className: 'orion-fusion-theme' }}>
         <DropdownList requestClose={() => {}}>
           <Item label="1st Option" onSelect={() => {}} />